Add L2 and cosine examples to verifying_mfd
Previous verifying_mfd lacked examples pertaining to strings and coordinates.
It also had no mention of the algorithm's capability of displaying clusters
that prevent MFD from holding.
This addition fixes lack of string and coordinates distance as well as adding
clusters, different parameter values and a brief explanation on how the MFD
works with different types of data on top of adding more more examples and
giving basic description of MFD.